backport the circular ref fix
