+++ packages/cli-repl/package.json
@@ -77,7 +77,7 @@
     "@mongosh/shell-evaluator": "^3.11.0",
     "@mongosh/snippet-manager": "^3.11.0",
     "@mongosh/types": "3.6.2",
-    "@mongodb-js/mongodb-ts-autocomplete": "^0.2.0",
+    "@mongodb-js/mongodb-ts-autocomplete": "^0.2.2",
     "@segment/analytics-node": "^1.3.0",
     "ansi-escape-sequences": "^5.1.2",
     "askcharacter": "^2.0.4",
